+++ v3_00/settings.py
@@ -65,6 +65,7 @@ SCENE_IMG = "crud_cafe_scene.png"
 
 # -- Faux OS Element Images --
 WINDOW_IMG = "window_blank_img.png" 
+WINDOW_BORDER_1_IMG = "window_blank_border_test_img.png" 
 WINDOW_HL_1_IMG = "window_blank_highlight_1_img.png" 
 WINDOW_HL_2_IMG = "window_blank_highlight_2_img.png" 
 WINDOW_SHELVED_1_IMG = "window_blank_shelved_test_img.png"